:art:Modify : 프로필페이지 컴포넌트 세부화
